move int_to_string_array to separate file, make it bash script
